README update
I updated the readme file with requirements for Android App development. I took the current version of my Android Studio (newer versions should be fine as well). As far as the SDK goes we don't use any newest features of Android, therefore it should be fairly easy just to change the target or minimum. This might change when we build in the background Service and Notification features.