Demo screens for menu items
When a button is pressed in the Menu it switches the content of the Navigation Drawer using fragments. Right now it only displays a message saying what menu item was pressed. Back button from Navigation Activity now minimises the App instead of going back to log in screen. To get back to log in screen sign out in the menu should be pressed.