working simulation, 5 pipes, 1 large pipe, 1 tank, 1 output, graphs
The simulation is now working, we have 5 pipes connected to 1 larger pipe. The larger pipe flows into the tank which has volume, level and temperature sensors. There is one output for the treated water. We need to add another valve to release water into the river in case the tank becomes full. We need to implement functions to generate data for different times of the day.

Co-Authored-By: tiarnan-g <32198808+tiarnan-g@users.noreply.github.com>
Co-Authored-By: shamashanf <8364724+shamashanf@users.noreply.github.com>