Refactoring and JavaDocs, readme edit
Simple package refactoring including the creation of main package (for the main class), GUI package (for the FXML files and the controller) and resources packages. I also wrote some simple JavaDoc documentation for methods and classes. It is not perfect, but it explains some basic functionality.

I also added the Java JDK in the requirements to use for the development (includes a link to download it).