UI design update, refactoring
UI CHANGES:
Notification demo scene now shows the percentage instead of the pure string data that is being sent over to Android.
Ability to launch Desktop app in full-screen with launch configuration dialog.
Minimum resolution set to 1280x720. It can be resized or full-screened, but could never be resized to be smaller than that.
Whole UI reworked to match the colours of the Android App. Blue and pink colour scheme. Also it is mainly based around GridPanes now instead of HBox or VBox.
Added CSS stylesheet for slider.

REFACTORING:
Buttons to change scenarios have separate functions.