Network check on desktop app, refactoring, android app settings
Desktop app checks whether the device is connected to a network. If it's not connected it launches an alert and prompts user to connect to a network and relaunch the app. Some little refactoring done as well. Settings activity created in the Android app. It is implemented as a Preference Fragment, which saves the IP address as a SharedPreference. Settings can then be read from anywhere in the app. IP is validated before actually saved to avoid crashes.