Slider animation, little refactoring
Slider now moves from 0 to 24 when START button pressed. When it reaches 24 it resets back to zero. The movement through can be stopped at any time. Implemented as a new Thread, so that the processing doesn't slow down the UI. Later all the file operations will be added to this thread.