Charts, file reading, refactoring
Added data reading from files produced by the simulation. The data is processed and displayed on the UI. There are two different charts, one for the main pipe and one for the tank capacity. Only one scenario is provided and notification system onn android has to be improved to reflect the new changes.